\name{birthProbs}
\alias{birthProbs}
\title{
Birth Probabilities
}
\description{
This function defines the probability of proposing an addition of a change-point. 
}
\usage{
birthProbs(LRange)
}
\arguments{
  \item{LRange}{ The range of possible values for the number of change-points.
}
}
\value{
  \item{probs}{Vector of probabilities}
}
\author{
Panagiotis Papastamoulis
}

