% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalibrationLasso.R
\name{CalibrationLasso}
\alias{CalibrationLasso}
\title{Function to compute $beta, the covariate balancing weights}
\usage{
CalibrationLasso(d, X, c = 1.1, maxIterPen = 100, PostLasso = F,
  trace = F, maxIter = 1000)
}
\arguments{
\item{d}{Treatment indicator.}

\item{X}{Matrix of covariates.}

\item{c}{Constant for the overall penalty level.}

\item{maxIterPen}{Maximal number of iterations for penalty estimation.}

\item{PostLasso}{if TRUE computes the Post-Lasso solution.}

\item{trace}{if TRUE print convergence info.}

\item{maxIter}{Maximal number of iteration for Lasso program.}
}
\value{
SHat Set of indices of non-zero elements in estimated beta.

betaLasso Lasso solution.

betaPL Post-Lasso solution.

lambda Overall penalty level.

psi Covariate-specific penalty loadings.

nbIter Number of iterations for penalty level estimation.

convergence 0 if convergence, -999 if not.
}
\description{
First step of the BEAST estimator. Uses the lbfgs function to perform L1-penalised minimization.
 A constant must be included as the first column in X.
 Last edited: 11 janvier 2016.
}

