\name{computePosteriorParameters}
\alias{computePosteriorParameters}
\title{
Compute empirical posterior parameters
}
\description{
Compute empirical posterior parameters
}
\usage{
computePosteriorParameters(myDataList, priorParameters)
}
\arguments{
  \item{myDataList}{
Observed data.
}
  \item{priorParameters}{
Prior parameters.
}
}
\value{
list of posterior parameters
}
\author{
Panagiotis Papastamoulis
}

