\name{computePosteriorParametersFree}
\alias{computePosteriorParametersFree}
\title{
Posterior parameters
}
\description{
Posterior parameters
}
\usage{
computePosteriorParametersFree(myDataList, priorParameters)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{myDataList}{
 observed data.
}
  \item{priorParameters}{
  list of prior parameters.
}
}
\value{
 list of posterior parameters.
}
\author{
Panagiotis Papastamoulis
}

