% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataSim.R
\name{DataSim}
\alias{DataSim}
\title{DGP for Monte-Carlo Experiment}
\usage{
DataSim(n = 2000, p = 50, Ry = 0.5, Rd = 0.2, Intercept = T,
  rho = 0.5)
}
\arguments{
\item{n}{Sample size.}

\item{p}{Number of Covariates.}

\item{Ry}{Desired R-squared for outcome equation.}

\item{Rd}{Desired R-squared for treatment equation.}

\item{Intercept}{If TRUE returns an intercept also.}

\item{rho}{Correlation coefficients between two adjacent covariates.}
}
\description{
Last edited: 8 fevrier 2016
}

