    \name{FungalGrowthDataset}
     \docType{data}
     \alias{FungalGrowthDataset}
     \title{Fungal Growth Dataset}
     \description{Time-series dataset with \eqn{N\times R \times T} growth levels for \eqn{R = 3} replicates of \eqn{N = 411} objects (mutants) measured every 10 minutes for \eqn{T = 289} time-points. See Papastamoulis et al (2017) for a detailed description.}
     \usage{FungalGrowthDataset}
     \format{Time-series data.}
     \keyword{datasets}
     \references{
           %% ~put references to the literature/web site here ~
Papastamoulis P., Furukawa T., van Rhijn N., Bromley M., Bignell E. and Rattray M. (2017). Bayesian detection of piecewise linear trends in replicated time-series with application to growth data modelling. arXiv:\bold{1709.06111} [stat.AP]
}

