% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImmunizedATT.R
\name{ImmunizedATT}
\alias{ImmunizedATT}
\title{Function to compute theta, the Immunized ATT}
\usage{
ImmunizedATT(y, d, X, beta, mu = rep(0, ncol(X)), Immunity = T)
}
\arguments{
\item{y}{Outcome variable.}

\item{d}{Treatment indicator.}

\item{X}{Matrix of covariates.}

\item{beta}{Estimated Calibration parameter (Step 1).}

\item{mu}{Estimated Orthogonality parameter (Step 2).}

\item{Immunity}{If TRUE computes the immunized ATT, simple plug-in otherwise.}
}
\value{
theta Immunized ATT estimate.

sigma Asymptotic standerd error.
}
\description{
Last edited: 16 fevrier 2016.
}

