% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LassoFISTA.R
\name{LassoFISTA}
\alias{LassoFISTA}
\title{Computes Lasso solution with FISTA}
\usage{
LassoFISTA(betaInit = rep(0, ncol(X)), y, X, W = rep(1, nrow(X)),
  nopen = NULL, lambda, psi = rep(1, ncol(X)), tol = 1e-08,
  maxIter = 1000, trace = F)
}
\arguments{
\item{betaInit}{Starting value for the coefficient. SHould be a vector of dim ncol(X).}

\item{y}{vector of the dependent variable, normalizing it is a good idea.}

\item{X}{Matrix of Covariates, quicker if normalized.}

\item{W}{Vector of weights for each observation in the Least Square Objective.}

\item{nopen}{Set of indices of variables that should not be penalized.}

\item{lambda}{Overall penalty parameter.}

\item{psi}{vector of penalty loadings for each variable.}

\item{tol}{Stopping criteion: difference between the value of objective function at two iterations.}

\item{maxIter}{Maximal number of iterations for the algorithm.}

\item{trace}{if TRUE print algortihm info}
}
\description{
Computes the Lasso solution using the FISTA method of Beck and Teboulle (2014).
The objective function is given as the mean squared plus lambda times the L1-norm.
Penalty-loadings for each coefficients are allowed.
}

