\name{logLikelihoodFullModel}
\alias{logLikelihoodFullModel}
\title{
Log-likelihood function.
}
\description{
Log-likelihood function.
}
\usage{
logLikelihoodFullModel(myData, cutPoints, theta, startPoint)
}
\arguments{
  \item{myData}{
data
}
  \item{cutPoints}{
change-points.
}
  \item{theta}{
means.
}
  \item{startPoint}{
optional integer at least equal to 2.
}
}
\value{
 log-likelihood value.
}
\author{
Panagiotis Papastamoulis
}

