% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OrthogonalityReg.R
\name{OrthogonalityReg}
\alias{OrthogonalityReg}
\title{Function to compute mu, the covariate balancing weights}
\usage{
OrthogonalityReg(y, d, X, beta, method = "WLSLasso", c = 1.1,
  nopenset = c(1), RescaleY = F, maxIterPen = 10000,
  maxIterLasso = 1e+06, tolLasso = 1e-06, PostLasso = F, trace = F)
}
\arguments{
\item{y}{Outcome variable (not normalized).}

\item{X}{Matrix of covariates.}

\item{beta}{Calibrating parameter estimate from the first step.}

\item{method}{One of "WLSLasso" or "LinearOutcome".}

\item{c}{Constant for the overall penalty level.}

\item{nopenset}{Set of indices that should not be penalized. Default is intercept penalized.}

\item{RescaleY}{if TRUE rescale variable y.}

\item{maxIterPen}{Maximal number of iterations for penalty estimation.}

\item{maxIterLasso}{Maximal number of iterations in Lasso procedure.}

\item{tolLasso}{Tolerance for stopping criterion in Lasso minimization.}

\item{PostLasso}{if TRUE computes the PostLasso solution.}

\item{trace}{if TRUE print convergence info.}
}
\value{
SHat Set of indices of non-zero elements in estimated mu.

muLasso Lasso solution.

muPL Post-Lasso solution.

lambda Overall penalty level.

psi Covariate-specific penalty loadings.

nbIter Number of iterations for penalty level estimation.

convergence 0 if convergence, -555 if not because of Lasso minimization, -999 if not because of penalty estimation.
}
\description{
Second step of the BEAST estimator. Uses the LassoFISTA function to perform L1-penalised minimization.
 A constant must be included as the first column in X.
 Last edited: 19 fevrier 2016.
}

