\name{simMultiIndNormInvGamma}
\alias{simMultiIndNormInvGamma}
\title{
Prior random numbers
}
\description{
Generation of mean values according to the prior
}
\usage{
simMultiIndNormInvGamma(mu, nu, alpha, beta)
}
\arguments{
  \item{mu}{
means
}
  \item{nu}{
precision parameter
}
  \item{alpha}{
prior parameter
}
  \item{beta}{
prior parameter
}
}
\value{
null
}
\author{
Panagiotis Papastamoulis
}

