\name{truncatedPoisson}
\alias{truncatedPoisson}
\title{
Truncated Poisson pdf
}
\description{
Probability density function of the truncated Poisson distribution.
}
\usage{
truncatedPoisson(Lmax = 20, gammaParameter = 1)
}
\arguments{
  \item{Lmax}{
Max number
}
  \item{gammaParameter}{
Location parameter.
}
}
\value{
  \item{logPrior }{Log-pdf values}
}
\author{
Panagiotis Papastamoulis
}

